refactor: 나의 입맛 프로필 리팩토링 (#67)
- 코드 리뷰 반영
- 배럴 사용
