+++ src/components/my-page/TasteTypeResult.tsx
@@ -4,7 +4,8 @@ import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 
 import { Button, RetryIcon } from '@/components'
-import { useTasteType, useUserTasteType } from '@/hooks/my-page/useTasteType'
+import { IMAGE_URLS } from '@/constants'
+import { useUserTasteType } from '@/hooks/my-page'
 import { useLoginStore } from '@/stores/useLoginStore'
 import { getUsername } from '@/utils/getUsername'
 
@@ -12,10 +13,8 @@ const TasteTypeResult = () => {
   const router = useRouter()
   const { user } = useLoginStore()
   const username = getUsername(user)
-  const { tasteType } = useTasteType()
   const { data: tasteTypeData } = useUserTasteType(user?.id || '')
-  const mainData = tasteTypeData?.data
-  const tasteTypeName = tasteType?.tasteType || '깔끔고소'
+  const tasteTypeName = tasteTypeData?.tasteType || '깔끔고소'
 
   return (
     <article className="flex-center flex-col">
@@ -26,7 +25,7 @@ const TasteTypeResult = () => {
         <div className="flex-center h-full flex-col gap-[10px]">
           <div className="flex-center h-20 w-20 rounded-full bg-red-600">
             <Image
-              src={mainData?.imageUrl || ''}
+              src={tasteTypeData?.imageUrl || IMAGE_URLS.MyPage.BaseTaste}
               alt="테스트 결과 이미지"
               width={60}
               height={53}
@@ -36,7 +35,7 @@ const TasteTypeResult = () => {
             {username}님 취향 유형은 &apos;{tasteTypeName}&apos; 유형입니다
           </h1>
           <p className="text-white-100 text-lg font-normal">
-            {mainData?.description?.join(' ')}
+            {tasteTypeData?.description?.join(' ')}
           </p>
           <Button
             variant="MY_PAGE_RETRY"
