+++ src/service/my-page/index.ts
@@ -1,0 +1,2 @@
+export { getUserAverageScores } from './gauge'
+export { getUserTasteType, type UserTasteTypeResult } from './tasteType'
